WORLD ECONOMIC ANALYSIS USING INNDICATORS
As analysis shows, the constant flow of data can obscure underlying trends. Decisions have to
be taken in a realistic way, unhindered by reading of economic data released almost daily that covers all
aspects of the economic sphere. But the sheer breadth of these releases and the underlying complexity
that they may not convey in the headline release means tools to help improve one’s understanding are
Required.