WEB TRAFFIC PREDICTION”
This internship is entitled “WEB TRAFFIC PREDICTION”. Now days, web traffic anticipating 
Is a significant issue as this can cause misfortunes in the activities of major sites. Time arrangement 
Topics have been an interesting issue for research. Anticipating future time arrangement esteems is 
one of the most troublesome issues in the business. The time arrangement field includes various 
issues, running from induction and examination to gauging and grouping. Estimating the 
organization traffic and updates continuously would be the most productive approach to pass on 
the data These days, we are excessively reliant on Google worker however in the event that we 
need to have a worker for huge clients we might have anticipated the quantity of clients from 
earlier years to stay away from worker breakdown. These days, web traffic anticipating is a 
significant issue as this can make misfortunes the activities of major sites. Time-arrangement 
gauging has been an interesting issue for research. Anticipating future time arrangement esteems 
is one of the most troublesome issues in the business. Estimating the organization's traffic and 
Updates continuously would be the most productive approach to pass on the data. In this project, 
We apply a forecasting model for predicting web traffic. Predicting web traffic can 
help web site owners in many ways, including: (a) determining an effective strategy for load 
balance of web pages remaining in the cloud; (b) forecasting future trends based on historical 
data and (c) understanding the user's behavior.