Image analytics
The problem statement for this project is to automate the process of quality analysis for an 
air compressor company using image classification. The company has a requirement to ensure that 
pictures of air compressors are taken and stored in the appropriate folders before the dispatch 
Currently, this process is manual and there is a possibility that some pictures may be missed. 
Therefore, the company wants to develop a system that can automatically detect and classify 
images of air compressors based on their category, such as make, model, and type. The proposed 
solution involves developing an image classification algorithm that can analyse images of air 
compressors and identify their category. The algorithm should be able to differentiate between 
different categories of images and ensure that all images of each category are captured and stored 
in the appropriate folder. The system should also be able to alert the quality analyst if any of the 
categories of images are missing, so that corrective action can be taken